# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_uthmanduro_BracketForge_internal_model.CreatePlayerRequest:
    properties:
      email:
        type: string
      metadata:
        items:
          type: integer
        type: array
      name:
        type: string
    required:
    - name
    type: object
...
  github_com_uthmanduro_BracketForge_internal_model.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  github_com_uthmanduro_BracketForge_internal_model.Organization:
    properties:
      created_at:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
  github_com_uthmanduro_BracketForge_internal_model.Player:
    properties:
      created_at:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      id:
        type: string
      metadata:
        items:
          type: integer
        type: array
      name:
        type: string
      organization_id:
        type: string
      updatedAt:
        type: string
    type: object
...
  github_com_uthmanduro_BracketForge_internal_model.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  internal_handler.CreateUpdateOrganizationRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@bracketforge.io
    name: BracketForge Support
  description: Racket sports tournament management engine. Supports single elimination,
    round robin, and group + knockout formats with set-based scoring, three-way tiebreakers,
    BYE handling, and result correction.
  termsOfService: http://swagger.io/terms/
  title: BracketForge API
  version: 1.0.0
paths:
  /organizations:
    post:
      consumes:
      - application/json
      description: Creates a new organisation. This is the onboarding entry point
        — no auth required.
      parameters:
      - description: Organisation name
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreateUpdateOrganizationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.SuccessResponse'
            - properties:
                ' Message':
                  type: string
                Data:
                  $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.Organization'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
      summary: Create organisation
      tags:
      - organisations
  /organizations/{id}:
    get:
      description: Returns the organisation for the authenticated user's org ID
      parameters:
      - description: Organisation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.SuccessResponse'
            - properties:
                ' Message':
                  type: string
                Data:
                  $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.Organization'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get organisation
      tags:
      - organisations
  /players:
    get:
      description: Returns all players in the authenticated organisation
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.SuccessResponse'
            - properties:
                ' Message':
                  type: string
                Data:
                  properties:
                    players:
                      items:
                        $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.Player'
                      type: array
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List players
      tags:
      - players
    post:
      consumes:
      - application/json
      description: Creates a new player profile within the organisation. Admin or
        organizer only.
      parameters:
      - description: Player details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.CreatePlayerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.SuccessResponse'
            - properties:
                ' Message':
                  type: string
                Data:
                  properties:
                    player_id:
                      type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create player
      tags:
      - players
  /players/{id}:
    delete:
      description: Permanently deletes a player from the organisation. Admin only.
      parameters:
      - description: Player ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete player
      tags:
      - players
    get:
      description: Returns a single player by ID
      parameters:
      - description: Player ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.SuccessResponse'
            - properties:
                ' Message':
                  type: string
                Data:
                  properties:
                    player:
                      $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.Player'
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get player
      tags:
      - players
    put:
      consumes:
      - application/json
      description: Updates name, email, or metadata for a player. Admin or organizer
        only.
      parameters:
      - description: Player ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated player details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.CreatePlayerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.SuccessResponse'
            - properties:
                ' Message':
                  type: string
                Data:
                  properties:
                    player:
                      $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.Player'
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update player
      tags:
      - players
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your JWT token as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"